backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Validation and serialization
pydantic==2.5.0

# Authentication (for future use)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client
aiohttp==3.9.1
httpx==0.25.2

# MQTT
paho-mqtt==1.6.1

# Kafka
aiokafka==0.8.11

# WebSocket
websockets==12.0

# Utilities
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1